Extract auth to separate file.
